requirements.txt
Cython
numpy
Pillow==8.1.1
tensorflow
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc
Flask==1.1.2
Flask-Reuploaded==0.3.2
Flask-RESTful
Flask-SQLAlchemy
gunicorn
flask-cors